more of Frame.__init__ can be made into frame processing
